formatting and file re-org
